backend/pyproject.toml
[project]
name = "woofbase"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "black>=25.11.0",
    "fastapi>=0.121.2",
    "httpx>=0.28.1",
    "pydantic>=2.12.4",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.44",
    "tenacity>=9.1.2",
    "uvicorn>=0.38.0",
    "psycopg2-binary>=2.9.0",
]
